+++ script/electron-builder.js
@@ -39,36 +39,90 @@ const builder = require("electron-builder")
 const Platform = builder.Platform
 
 
-const pngIcon = 'resources/app-icons/nightly/png/1024.png'
-const icoIcon = 'resources/app-icons/nightly/pulsar.ico'
+const pngIcon = 'resources/app-icons/beta.png'
+const icoIcon = 'resources/app-icons/beta.ico'
 
 let options = {
   "appId": "com.pulsar-edit.pulsar",
   "npmRebuild": false,
   "publish": null,
   files: [
+    // --- Inclusions ---
+    // Core Repo Inclusions
     "package.json",
-    "docs/**/*",
     "dot-atom/**/*",
     "exports/**/*",
-    "keymaps/**/*",
-    "menus/**/*",
-    "node_modules/**/*",
     "resources/**/*",
-    "script/**/*",
     "src/**/*",
     "static/**/*",
     "vendor/**/*",
+    "node_modules/**/*",
+
+    // Core Repo Test Inclusions
+    "spec/jasmine-test-runner.js",
+    "spec/spec-helper.js",
+    "spec/jasmine-junit-reporter.js",
+    "spec/spec-helper-functions.js",
+    "spec/atom-reporter.js",
+    "spec/jasmine-list-reporter.js",
+
+    // --- Exclusions ---
+    // Core Repo Exclusions
+    "!docs/",
+    "!keymaps/",
+    "!menus/",
+    "!script/",
+
+    // Git Related Exclusions
+    "!**/{.git,.gitignore,.gitattributes,.git-keep,.github}",
+    "!**/{.eslintignore,PULL_REQUEST_TEMPLATE.md,ISSUE_TEMPLATE.md,CONTRIBUTING.md,SECURITY.md}",
+
+    // Development Tools Exclusions
+    "!**/{npm-debug.log,yarn.lock,.yarn-integrity,.yarn-metadata.json,.npmignore}",
+    "!**/npm/{doc,html,man}",
+    "!.editorconfig",
+    "!**/{appveyor.yml,.travis.yml,circle.yml}",
+    "!**/{__pycache__,thumbs.db,.flowconfig,.idea,.vs,.nyc_output}",
+    "!**/*.{iml,o,hprof,orig,pyc,pyo,rbc,swp,csproj,sln,xproj}",
+    "!**/{.jshintrc,.pairs,.lint,.lintignore,.eslintrc,.jshintignore}",
+    "!**/{.coffeelintignore,.editorconfig,.nycrc,.coffeelint.json,.vscode,coffeelint.json}",
+
+    // Common File Exclusions
+    "!**/{.DS_Store,.hg,.svn,CVS,RCS,SCCS}",
+
+    // Build Chain Exclusions
+    "!**/*.{cc,h}", // Ignore *.cc and *.h files from native modules
+    "!**/*.js.map",
+    "!**/{Makefile}",
+    "!**/build/{binding.Makefile,config.gypi,gyp-mac-tool,Makefile}",
+    "!**/build/Release/{obj.target,obj,.deps}",
+
+    // Test Exclusions
+    "!**/pegjs/examples",
     "!**/node_modules/*/{test,__tests__,tests,powered-test,example,examples}",
+    "!**/node_modules/babel-core/lib/transformation/transforers/spec", // Ignore babel-core spec
+    "!**/{oniguruma,dev-live-reload,deprecation-cop,one-dark-ui,incompatible-packages,git-diff,line-ending-selector}/spec",
+    "!**/{link,grammar-selector,json-schema-traverse,exception-reporting,one-light-ui,autoflow,about,go-to-line,sylvester,apparatus}/spec",
+    "!**/{archive-view,autocomplete-plus,autocomplete-atom-api,autocomplete-css,autosave}/spec",
+
+    // Other Exclusions
+    "!**/._*",
     "!**/node_modules/*.d.ts",
     "!**/node_modules/.bin",
-    "!**/*.{iml,o,hprof,orig,pyc,pyo,rbc,swp,csproj,sln,xproj}",
-    "!.editorconfig",
-    "!**/._*",
-    "!**/{.DS_Store,.git,.hg,.svn,CVS,RCS,SCCS,.gitignore,.gitattributes}",
-    "!**/{__pycache__,thumbs.db,.flowconfig,.idea,.vs,.nyc_output}",
-    "!**/{appveyor.yml,.travis.yml,circle.yml}",
-    "!**/{npm-debug.log,yarn.lock,.yarn-integrity,.yarn-metadata.json}"
+    "!**/node_modules/native-mate",
+    "!node_modules/fuzzy-native/node_modules", // node_modules of the fuzzy-native package are only required for building it
+    "!**/node_modules/spellchecker/vendor/hunspell/.*",
+    "!**/git-utils/deps",
+    "!**/oniguruma/deps",
+    "!**/less/dist",
+    "!**/get-parameter-names/node_modules/testla",
+    "!**/get-parameter-names/node_modules/.bin/testla",
+    "!**/jasmine-reporters/ext",
+    "!**/deps/libgit2",
+    // These are only required in dev-mode, when pegjs grammars aren't precompiled
+      // "!node_modules/loophole", // Note: We do need these packages. Because our PegJS files _aren't_ all pre-compiled.
+      // "!node_modules/pegjs",    // Note: if these files are excluded, 'snippets' package breaks.
+      // "!node_modules/.bin/pegjs", // Note: https://github.com/pulsar-edit/pulsar/pull/206
   ],
   "extraResources": [
     {
@@ -95,7 +149,8 @@ let options = {
     "target": [
       { target: "appimage" },
       { target: "deb" },
-      { target: "rpm" }
+      { target: "rpm" },
+      { target: "tar.gz" }
     ],
   },
   "mac": {
@@ -111,7 +166,12 @@ let options = {
   },
   "extraMetadata": {
   },
-  "asarUnpack": ["node_modules/github/bin/*"]
+  "asarUnpack": [
+    "node_modules/github/bin/*",
+    "node_modules/github/lib/*", // Resolves Error in console
+    "**/node_modules/dugite/git/**", // Include dugite postInstall output (matching glob used for Atom)
+    "**/node_modules/spellchecker/**", // Matching Atom Glob
+  ]
 
 }